+++ include/boost/beast/core/detail/impl/read.hpp
@@ -13,7 +13,6 @@
 #include <boost/beast/core/bind_handler.hpp>
 #include <boost/beast/core/async_op_base.hpp>
 #include <boost/beast/core/flat_static_buffer.hpp>
-#include <boost/beast/core/stream_traits.hpp>
 #include <boost/asio/basic_stream_socket.hpp>
 #include <boost/asio/coroutine.hpp>
 #include <boost/throw_exception.hpp>
