+++ test/beast/core/stream_traits.cpp
@@ -13,6 +13,7 @@
 #include <boost/beast/_experimental/unit_test/suite.hpp>
 #include <boost/beast/core/error.hpp>
 #include <boost/beast/core/type_traits.hpp>
+#include <boost/asio/io_context.hpp>
 
 namespace boost {
 namespace beast {
@@ -161,7 +162,7 @@ public:
     };
     
     void
-    testJavadoc()
+    testGetLowestLayerJavadoc()
     {
         write_stream<without> s;
         BOOST_STATIC_ASSERT(
@@ -179,10 +180,101 @@ public:
     //--------------------------------------------------------------------------
 
     void
+    testExecutorType()
+    {
+    }
+
+    void
+    testExecutorTypeJavadoc()
+    {
+    }
+
+    //--------------------------------------------------------------------------
+
+    struct sync_read_stream
+    {
+        template<class MutableBufferSequence>
+        std::size_t
+        read_some(MutableBufferSequence const&);
+
+        template<class MutableBufferSequence>
+        std::size_t
+        read_some(MutableBufferSequence const&,
+            error_code& ec);
+    };
+
+    struct sync_write_stream
+    {
+        template<class ConstBufferSequence>
+        std::size_t
+        write_some(ConstBufferSequence const&);
+
+        template<class ConstBufferSequence>
+        std::size_t
+        write_some(
+            ConstBufferSequence const&, error_code&);
+    };
+
+    struct async_read_stream
+    {
+        net::io_context::executor_type
+        get_executor();
+
+        template<class MutableBufferSequence, class ReadHandler>
+        void
+        async_read_some(
+            MutableBufferSequence const&, ReadHandler&&);
+    };
+
+    struct async_write_stream
+    {
+        net::io_context::executor_type
+        get_executor();
+
+        template<class ConstBufferSequence, class WriteHandler>
+        void
+        async_write_some(
+            ConstBufferSequence const&, WriteHandler&&);
+    };
+
+    struct sync_stream : sync_read_stream, sync_write_stream
+    {
+    };
+
+    struct async_stream : async_read_stream, async_write_stream
+    {
+        using async_read_stream::get_executor;
+    };
+
+    BOOST_STATIC_ASSERT(is_sync_read_stream<sync_read_stream>::value);
+    BOOST_STATIC_ASSERT(is_sync_write_stream<sync_write_stream>::value);
+    BOOST_STATIC_ASSERT(is_sync_read_stream<sync_stream>::value);
+    BOOST_STATIC_ASSERT(is_sync_write_stream<sync_stream>::value);
+    BOOST_STATIC_ASSERT(is_sync_stream<sync_stream>::value);
+
+    BOOST_STATIC_ASSERT(is_async_read_stream<async_read_stream>::value);
+    BOOST_STATIC_ASSERT(is_async_write_stream<async_write_stream>::value);
+    BOOST_STATIC_ASSERT(is_async_read_stream<async_stream>::value);
+    BOOST_STATIC_ASSERT(is_async_write_stream<async_stream>::value);
+    BOOST_STATIC_ASSERT(is_async_stream<async_stream>::value);
+
+    BOOST_STATIC_ASSERT(! is_sync_read_stream<sync_write_stream>::value);
+    BOOST_STATIC_ASSERT(! is_sync_write_stream<sync_read_stream>::value);
+    BOOST_STATIC_ASSERT(! is_async_read_stream<async_write_stream>::value);
+    BOOST_STATIC_ASSERT(! is_async_write_stream<async_read_stream>::value);
+
+    BOOST_STATIC_ASSERT(! is_sync_stream<async_stream>::value);
+    BOOST_STATIC_ASSERT(! is_async_stream<sync_stream>::value);
+
+    //--------------------------------------------------------------------------
+
+    void
     run() override
     {
         testGetLowestLayer();
-        testJavadoc();
+        testGetLowestLayerJavadoc();
+        testExecutorType();
+        testExecutorTypeJavadoc();
     }
 };
