+++ test/beast/core/type_traits.cpp
@@ -79,80 +79,5 @@ struct H
 BOOST_STATIC_ASSERT(is_completion_handler<H, void(int)>::value);
 BOOST_STATIC_ASSERT(! is_completion_handler<H, void(void)>::value);
 
-//
-// stream concepts
-//
-
-struct sync_write_stream
-{
-    net::io_context&
-    get_io_service();
-
-    template<class ConstBufferSequence>
-    std::size_t
-    write_some(ConstBufferSequence const& buffers);
-
-    template<class ConstBufferSequence>
-    std::size_t
-    write_some(
-        ConstBufferSequence const& buffers, error_code& ec);
-};
-
-struct sync_read_stream
-{
-    template<class MutableBufferSequence>
-    std::size_t
-    read_some(MutableBufferSequence const& buffers);
-
-    template<class MutableBufferSequence>
-    std::size_t
-    read_some(MutableBufferSequence const& buffers,
-        error_code& ec);
-};
-
-struct sync_stream : sync_read_stream, sync_write_stream
-{
-};
-
-BOOST_STATIC_ASSERT(! is_sync_read_stream<sync_write_stream>::value);
-BOOST_STATIC_ASSERT(! is_sync_write_stream<sync_read_stream>::value);
-
-BOOST_STATIC_ASSERT(is_sync_read_stream<sync_read_stream>::value);
-BOOST_STATIC_ASSERT(is_sync_write_stream<sync_write_stream>::value);
-
-BOOST_STATIC_ASSERT(is_sync_read_stream<sync_stream>::value);
-BOOST_STATIC_ASSERT(is_sync_write_stream<sync_stream>::value);
-
-namespace {
-
-using stream_type = net::ip::tcp::socket;
-
-struct not_a_stream
-{
-    void
-    get_io_service();
-};
-
-BOOST_STATIC_ASSERT(has_get_executor<stream_type>::value);
-BOOST_STATIC_ASSERT(is_async_read_stream<stream_type>::value);
-BOOST_STATIC_ASSERT(is_async_write_stream<stream_type>::value);
-BOOST_STATIC_ASSERT(is_async_stream<stream_type>::value);
-BOOST_STATIC_ASSERT(is_sync_read_stream<stream_type>::value);
-BOOST_STATIC_ASSERT(is_sync_write_stream<stream_type>::value);
-BOOST_STATIC_ASSERT(is_sync_stream<stream_type>::value);
-
-BOOST_STATIC_ASSERT(! has_get_executor<not_a_stream>::value);
-BOOST_STATIC_ASSERT(! is_async_read_stream<not_a_stream>::value);
-BOOST_STATIC_ASSERT(! is_async_write_stream<not_a_stream>::value);
-BOOST_STATIC_ASSERT(! is_sync_read_stream<not_a_stream>::value);
-BOOST_STATIC_ASSERT(! is_sync_write_stream<not_a_stream>::value);
-
-BOOST_STATIC_ASSERT(is_sync_read_stream<test::stream>::value);
-BOOST_STATIC_ASSERT(is_sync_write_stream<test::stream>::value);
-BOOST_STATIC_ASSERT(is_async_read_stream<test::stream>::value);
-BOOST_STATIC_ASSERT(is_async_write_stream<test::stream>::value);
-
-} // (anonymous)
-
 } // beast
 } // boost
